update filter xe cho thue va json filter

+++ BE/src/main/java/com/mycar/nhom13/Repository/CarRepository.java
@@ -13,7 +13,7 @@ import org.springframework.data.repository.query.Param;
 
 public interface CarRepository extends JpaRepository<Car,Integer>, JpaSpecificationExecutor<Car>{
     
-    Car findByCarId(int id);
+    Car findByCarId(Long id);
     
     Page<Car> findByStatus(String status, Pageable pageable);
